% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Function to add two numbers}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{a}{First number}

\item{b}{Second number}
}
\value{

}
\description{
Function to add two numbers
}
\examples{
add_two_numbers(2, 4)
add_two_numbers(7, -4)
}
