% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.r
\name{calculation_1}
\alias{calculation_1}
\title{Function to return (a - b)*(a + b) using a and b as inputs.}
\usage{
calculation_1(a, b)
}
\arguments{
\item{a}{numeric value}

\item{b}{a numeric value added after b}
}
\value{
return The output of the function
}
\description{
Function to return (a - b)*(a + b) using a and b as inputs.
}
\examples{
calculation_1(10,5)
calculation_1(8,3)
}
