% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.r
\name{calculation_2}
\alias{calculation_2}
\title{Function to return (a + b)*(a + b)
Takes in numeric a and b as input.}
\usage{
calculation_2(a, b)
}
\arguments{
\item{a}{numeric value}

\item{b}{a numeric value added after b}
}
\value{
The output of the function
}
\description{
Function to return (a + b)*(a + b)
Takes in numeric a and b as input.
}
\examples{
calculation_2(2,3)
calculation_2(5,8)
}
