% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Celsuis to Fahrenheit
Function converts celsius temperatures to fahrenheit temperatures}
\usage{
celsius_to_fahrenheit(temp_C)
}
\arguments{
\item{temp_C}{}
}
\value{
The temperature after being converted to fahrenheit
}
\description{
Celsuis to Fahrenheit
Function converts celsius temperatures to fahrenheit temperatures
}
\examples{
celsius_to_fahrenheit(40)
celsius_to_fahrenheit(-3)
}
