% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cube}
\alias{cube}
\title{Cube function}
\usage{
cube(x)
}
\arguments{
\item{x}{A number}
}
\value{
Cube of x which is x^3
}
\description{
This function calculates the cube of the number provided as an input to the function.
}
\examples{
cube(2)
cube(0.1)

}
