% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_name}
\alias{hello_name}
\title{Hello function}
\usage{
hello_name(name)
}
\arguments{
\item{name}{A name.}
}
\value{
A greating to \code{name}.
}
\description{
Returns a greeting to the name passed as a parameter
}
\examples{
hello_name("Mary")
}
