% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{multiply_two_numbers}
\alias{multiply_two_numbers}
\title{Function to multiply two numbers}
\usage{
multiply_two_numbers(a, b)
}
\arguments{
\item{a}{First Number}

\item{b}{Second Number}
}
\value{

}
\description{
Function to multiply two numbers
}
\examples{
multiply_two_numbers(2, 4)
multiply_two_numbers(7, -4)
}
