% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{number_square}
\alias{number_square}
\title{Calculate square of a number}
\usage{
number_square(x)
}
\arguments{
\item{x}{should be a number not text}
}
\value{
the function will return the square of the argument.
}
\description{
This function returns the square of an number given as argument
}
\examples{
number_square(5)
}
