% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/plus_four.R
\name{plus_four}
\alias{plus_four}
\title{plus_four}
\usage{
plus_four(x)

plus_four(x)
}
\arguments{
\item{x}{a number}
}
\value{
x plus four

x plus four
}
\description{
plus_four

plus_four
}
\examples{
plus_four(4)
plus_four(8)

plus_four(4)
plus_four(8)

}
