% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plus_three}
\alias{plus_three}
\title{Add 3 to number}
\usage{
plus_three(x)
}
\arguments{
\item{x}{should be a number not text}
}
\value{
the function will return a number with 3 added to argument.
}
\description{
This function adds 3 to any given number by taking the number as an argument
}
\examples{
plus_three(5)
}
