% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square}
\alias{square}
\title{Square function}
\usage{
square(x)
}
\arguments{
\item{x}{A number}
}
\value{
Square of x which is x^2
}
\description{
This function calculates the square of the number provided as an input to the function.
}
\examples{
square(10)
square(4)

}
