% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{square_it}
\alias{square_it}
\title{Squared of a number}
\usage{
square_it(num)
}
\arguments{
\item{num}{A number.}
}
\value{
The value of \code{num} squared.
}
\description{
Returns the squared value of the number passed as a parameter.
}
\examples{
square_it(4)
}
