% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{squared}
\alias{squared}
\title{squared
This function prints the square of a value, multiplying the value by itself}
\usage{
squared(val)
}
\arguments{
\item{val}{An integer that you want to square}
}
\value{
This function does not return any data, but instead prints the square of a value
}
\description{
squared
This function prints the square of a value, multiplying the value by itself
}
\examples{
squared(2)
}
