% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sum_f}
\alias{sum_f}
\title{Adds two numbers}
\usage{
sum_f(x, y)
}
\arguments{
\item{x}{A number}
\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
The function returns the sum of two numbers
}
\examples{
sum_f(2,3)
}