% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Function to add two numbers}
\usage{
add_two_numbers(a, b)
}
\arguments{
\item{b}{}
}
\value{

}
\description{
Function to add two numbers
}
\examples{
add_two_numbers(1, 3)
add_two_numbers(-5, 6)
}
