% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication.R
\name{multiply_two_numbers}
\alias{multiply_two_numbers}
\title{Function to multiply two numbers}
\usage{
multiply_two_numbers(a, b)
}
\arguments{
\item{a}{first number}

\item{b}{second number}
}
\value{

}
\description{
Function to multiply two numbers
}
\examples{
multiply_two_numbers(4, 5)
multiply_two_numbers(6, -3)
}
