% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{product_f}
\alias{product_f}
\title{Multiplies two numbers}
\usage{
product_f(x, y)
}
\arguments{
\item{x}{A number}
\item{y}{A number}
}
\value{
The product of \code{x} and \code{y}
}
\description{
The function returns the product of two numbers
}
\examples{
product_f(2,3)
}